+++ stimela/__init__.py
@@ -16,13 +16,10 @@ from recipe import Recipe as Pipeline
 from recipe import Recipe, PipelineException
 from stimela import docker
 
-from stimela.utils import stimela_logger
+from stimela.utils import logger
 
 LOG_HOME = os.path.expanduser("~/.stimela")
-LOG_IMAGES = LOG_HOME + "/stimela_images.log"
-LOG_CONTAINERS = LOG_HOME + "/stimela_containers.log"
-LOG_PROCESS = LOG_HOME + "/stimela_process.log"
-LOG_CABS = LOG_HOME + "/stimela_cab.log"
+LOG_FILE = LOG_HOME + "/stimela_logfile.json"
 
 
 BASE = os.listdir(cargo.BASE_PATH)
@@ -83,6 +80,7 @@ def build(argv):
             help="Do not use cache when building the image")
 
     args = parser.parse_args(argv)
+    log = logger.StimelaLogger(LOG_FILE)
 
     if args.base:
         for image in BASE:
@@ -90,6 +88,10 @@ def build(argv):
             image = "stimela/{0}:{1}".format(image, __version__)
             docker.build(image,
                          dockerfile)
+
+        log.log_image(image, replace=True)
+        log.write()
+
         return 0
 
     workdir = "/home/{}/output/".format(USER)
@@ -100,6 +102,7 @@ def build(argv):
                   "USER {:s}".format(USER)]
 
     no_cache = ["--no-cache"] if args.no_cache else []
+
 
     if args.cab:
         cab_args = args.cab.split(",")
@@ -115,14 +118,9 @@ def build(argv):
                      path,
                      build_args=build_args, args=no_cache)
 
-        img = stimela_logger.Image(LOG_CABS)
-        img.add(dict(name=cab))
-        img.write()
+        log.log_image(image, replace=True, cab=True)
+        log.write()
         return
-
-    # clear old cabs
-    img = stimela_logger.Image(LOG_CABS)
-    img.clear()
 
     for image in CAB:
         IGNORE = args.ignore_cabs.split(",")
@@ -135,9 +133,9 @@ def build(argv):
                      dockerfile,
                      build_args=build_args, args=no_cache)
 
-        img.add(dict(name=image))
+        log.log_image(image, replace=True, cab=True)
+        log.write()
 
-    img.write()
 
 
 def info(cabname, header=False):
@@ -206,7 +204,6 @@ def run(argv):
     args = parser.parse_args(argv)
     tag =  None
 
-
     _globals = dict(STIMELA_INPUT=args.input, STIMELA_OUTPUT=args.output,
                     STIMELA_MSDIR=args.msdir,
                     CAB_TAG=tag)
@@ -248,14 +245,14 @@ def pull(argv):
             help="Tag")
 
     args = parser.parse_args(argv)
+    log = logger.StimelaLogger(LOG_FILE)
 
     if args.image:
         for image in args.image:
-            img = stimela_logger.Image(LOG_IMAGES)
 
             if not img.find(image):
                 docker.pull(image)
-                img.add(dict(name=image, tag=tagargs.tag))
+                log.log_image(image)
     else:
 
         base = []
@@ -266,11 +263,11 @@ def pull(argv):
         base = set(base)
 
         for image in base:
-            img = stimela_logger.Image(LOG_IMAGES)
-
-            if not img.find(image) and image not in ["stimela/ddfacet", "radioastro/ddfacet"]:
+            if image not in ["stimela/ddfacet", "radioastro/ddfacet"]:
                 docker.pull(image)
-                img.add(dict(name=image, tag=args.tag))
+                log.log_image(image)
+
+    log.write()
 
 
 def images(argv):
@@ -282,15 +279,16 @@ def images(argv):
     add = parser.add_argument
 
     add("-c", "--clear", action="store_true",
-            help="Clear images log file")
+            help="Clear the logfile that keeps track of stimela images. This does not do anythig to the images.")
 
     args = parser.parse_args(argv)
 
-    img = stimela_logger.Image(stimela.LOG_IMAGES)
-    img.display()
+    log = logger.StimelaLogger(LOG_FILE)
+    log.display('images')
 
     if args.clear:
-        img.clear()
+        log.clear('images')
+        log.write()
 
 
 def containers(argv):
@@ -302,14 +300,15 @@ def containers(argv):
     add = parser.add_argument
 
     add("-c", "--clear", action="store_true",
-            help="Clear containers log file")
+            help="Clear the log file that keeps track of stimela containers. This doesn't do anything to the containers.")
 
     args = parser.parse_args(argv)
 
-    conts = stimela_logger.Container(stimela.LOG_CONTAINERS)
-    conts.display()
+    log = logger.StimelaLogger(LOG_FILE)
+    log.display('containers')
     if args.clear:
-        conts.clear()
+        log.clear('containers')
+        log.write()
 
 
 def ps(argv):
@@ -321,14 +320,15 @@ def ps(argv):
     add = parser.add_argument
 
     add("-c", "--clear", action="store_true",
-            help="Clear Log file")
+            help="Clear logfile that keeps track of stimela processes. This doesn't do anything ot the processes themselves.")
 
     args = parser.parse_args(argv)
 
-    procs = stimela_logger.Process(stimela.LOG_PROCESS)
-    procs.display()
+    log = logger.StimelaLogger(LOG_FILE)
+    log.display('processes')
     if args.clear:
-        procs.clear()
+        log.clear('processes')
+        log.write()
 
 
 def kill(argv):
@@ -344,42 +344,96 @@ def kill(argv):
 
     args = parser.parse_args(argv)
 
-    procs = stimela_logger.Process(LOG_PROCESS)
+    log = logger.StimelaLogger(LOG_FILE)
 
-    for pid in map(int, args.pid):
+    for pid in args.pid:
 
-        found = procs.find(pid)
+        found = pid in log.info['processes'].keys()
 
         if not found:
-            print "Could not find process {0}".format(pid)
+            print("Could not find process {0}".format(pid))
             continue
 
-        conts = stimela_logger.Container(LOG_CONTAINERS)
-        lines = []
-
-        procs.rm(pid)
-        procs.write()
-
-        for cont in conts.lines:
-            if cont.find(str(pid)) > 0:
-                lines.append(cont)
-
-        for line in lines:
-            cont, _id, utime, _pid, status = line.split()
-            if status.find("removed")<0:
-                cont_ = docker.Container(None, cont, None, None)
-                cont_.started = True
-                cont_.stop()
-                cont_.remove()
-
-                conts.rm(line)
-
-        conts.write()
-
         try:
-            os.kill(pid, signal.SIGKILL)
+            os.kill(int(pid), signal.SIGINT)
         except OSError:
-            pass
+            raise OSError('Process with PID {} could not be killed'.format(pid))
+
+        log.remove('processes', pid)
+    log.write()
+
+
+def clean(argv):
+    for i, arg in enumerate(argv):
+        if (arg[0] == '-') and arg[1].isdigit(): argv[i] = ' ' + arg
+
+    parser = ArgumentParser(description='Convience tools for cleaning up after stimela')
+    add = parser.add_argument
+
+    add("-ai", "--all-images", action="store_true",
+        help="Remove all images pulled/built by stimela. This include CAB images")
+    
+    add("-ab", "--all-base", action="store_true",
+        help="Remove all base images")
+
+    add("-ac", "--all-cabs", action="store_true", 
+        help="Remove all CAB images")
+
+    add("-aC", "--all-containers", action="store_true",
+        help="Stop and/or Remove all stimela containers")
+
+    args = parser.parse_args(argv)
+
+    log = logger.StimelaLogger(LOG_FILE)
+
+    if args.all_images:
+        images = log.info['images'].keys()
+        for image in images:
+            utils.xrun('docker', ['rmi', image])
+            log.remove('images', image)
+            log.write()
+
+    if args.all_base:
+        images = log.info['images'].keys()
+        for image in images:
+            if log.info['images'][image]['CAB'] is False:
+                utils.xrun('docker', ['rmi', image])
+                log.remove('images', image)
+                log.write()
+
+    if args.all_cabs:
+        images = log.info['images'].keys()
+        for image in images:
+            if log.info['images'][image]['CAB']:
+                utils.xrun('docker', ['rmi', image])
+                log.remove('images', image)
+                log.write()
+
+        
+    if args.all_containers:
+        containers = log.info['containers'].keys()
+        for container in containers:
+            cont = docker.Container(log.info['containers'][container]['IMAGE'], container)
+            try:
+                status = cont.info()['State']['Status']
+            except OSError:
+                print('Could not inspect container {}. It probably doesn\'t exist, will remove it from log'.format(container))
+                status = False
+                log.remove('containers', container)
+                log.write()
+                continue
+
+            if status == 'running':
+            # Kill the container instead of stopping it, so that effect can be felt py parent process
+                utils.xrun('docker', ['kill', container])
+                cont.remove()
+                log.remove('containers', container)
+                log.write()
+            elif status == 'exited':
+                cont.remove()
+                log.remove('containers', container)
+                log.write()
+
 
 def main(argv):
     for i, arg in enumerate(argv):
@@ -404,7 +458,8 @@ def main(argv):
     options = []
     commands = dict(pull=pull, build=build, run=run,
                     images=images, cabs=cabs, ps=ps,
-                    containers=containers, kill=kill)
+                    containers=containers, kill=kill,
+                    clean=clean)
 
     command = "failure"
 
@@ -438,6 +493,7 @@ images  : List stimela images
 cabs    : List active stimela containers
 ps      : List running stimela scripts
 kill    : Gracefully kill runing stimela process
+clean   : Clean up tools for stimela
 
 """)