+++ CHANGES.md
@@ -13,3 +13,11 @@
 * Add lwimager, wslcean, casa imagers (instead of a single imager image)
 * Add owlcat base image
 * Add container, docker image, and job loging
+
+# 0.2.5 
+- Improve hnadling of ctrl+c. Send terminate
+  signal to container.   
+- Improve handling of FITS images when predicting
+  visibilities; allow for minimal FITS images.  
+- Add clean fuction to help clean up after stimela  
+- Improve/Fx logging of images and containers